WIP: agent resolver for checkpoint
This adds an agent resolver that:

- expects to receive an Actor that has a #request method that returns a
Rack request, which then gets passed to each attribute generator
(currently, just InstitutionFinder)

- delegates to Keycard::RequestAttributes to call all the attribute
generators (currently, just InstitutionFinder - but this will be the
hook where we add additional attribute generators

- unpacks the attributes and uses the given agent_factory to turn the
attributes into agents.

Currently there is no default agent_factory, but we expect to delegate
to Checkpoint to do this, but there's no integration test for that yet -
hence work in progress.